addons: Reorganize module
